odoo_stackit/requirement.txt
django-cors-headers
django-tinymce4-lite